Refactor under responsive-toolbar-group
