+++ client/my-sites/plugins/categories/responsive-toolbar-group/style.scss
@@ -1,4 +1,4 @@
-.responsive-toolbar-group {
+.responsive-toolbar-group__dropdown {
 	position: relative;
 	width: 100%;
 
@@ -35,27 +35,49 @@
 		left: 0;
 		right: 0;
 		z-index: -1;
-
-		// for testing purposes only
-		// z-index: 2 !important
-		// height: auto !important;
-		// filter: alpha(opacity=50);
-		// opacity:0.5;
 	}
-		
-		.responsive-toolbar-group__button-item {
-			flex-basis: max-content;
-			white-space: nowrap;
-			z-index: 0;
-		}
 
-		.responsive-toolbar-group__button-item, .responsive-toolbar-group__menu-item {
-			&:focus {
-				&::before {
-					// disable blue box on focus
-					box-shadow: none !important;
-					outline: none !important;
-				}
+	.responsive-toolbar-group__button-item {
+		flex-basis: max-content;
+		white-space: nowrap;
+		z-index: 0;
+	}
+
+	.responsive-toolbar-group__button-item, .responsive-toolbar-group__menu-item {
+		&:focus {
+			&::before {
+				// disable blue box on focus
+				box-shadow: none !important;
+				outline: none !important;
 			}
 		}
 	}
+}
+
+.responsive-toolbar-group__swipe {
+	width: 100%;
+
+	.responsive-toolbar-group__swipe-list {
+		padding: 0 8px;
+		display: flex;
+		flex-wrap: nowrap;
+		overflow-x: scroll;
+		overflow-y: hidden;
+		border: none;
+
+		.responsive-toolbar-group__swipe-item {
+			font-size: 0.875rem;
+		}
+
+		// Core override - prevent buttons from wordwrapping content
+		>div {
+			flex-shrink: 0;
+		}
+	}
+
+	// Remove on-focus, on-click border
+	.components-toolbar .components-button::before {
+		box-shadow: none;
+	}
+
+}
